Split utils.js into two files
